app_api/pyproject.toml
[project]
name = "app-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.135.1",
    "pandas>=2.2.2,<3.0.0",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
# Ce groupe permet à la racine de récupérer les bibliothèques de l'API
app-api = [
    "fastapi>=0.135.1",
    "pandas>=2.2.2,<3.0.0",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
]

[tool.pytest.ini_options]
pythonpath = ["app_api"]
testpaths = ["../tests"]
